(index) Route position bitmask from converter to index
This commit changes the term metadata column from a compressed byte array to a long array.  Each long consists of a byte containing the old term metadata, and then a 56 bit mask containing approximate position data.

At this point, the index constructor just extracts the term metadata byte and proceeds with the original code.

+++ code/index/test/nu/marginalia/index/CombinedIndexReaderTest.java
@@ -282,7 +282,7 @@ public class CombinedIndexReaderTest {
                 var meta = metaByDoc.get(doc);
 
                 List<String> keywords = words.stream().map(w -> w.keyword).toList();
-                byte[] metadata = new byte[words.size()];
+                long[] metadata = new long[words.size()];
                 for (int i = 0; i < words.size(); i++) {
                     metadata[i] = words.get(i).termMetadata;
                 }
